[WIP] Separate c-TPE procs
